feat: 쪽지 목록 컴포넌트 분리
